Revert "Rename thumbnail test to media_msg so that it can cover blurhash too"
This reverts commit c55374cf7cfd61fbee27573ce4d4a86f958818a2.
