+++ native/test/src/tests.rs
@@ -4,7 +4,6 @@ mod calendar;
 mod categories;
 mod formatted_body;
 mod invitation;
-mod media_msg;
 mod msg_edit;
 mod news;
 mod onboarding;
@@ -19,5 +18,6 @@ mod spaces;
 mod super_invites;
 mod tasks;
 mod templates;
+mod thumbnail;
 mod typing;
 mod verification;
