+++ native/test/src/tests/thumbnail.rs
